+++ site/site.config.mjs
@@ -23,7 +23,7 @@ export default {
           name: 'used',
           meta: { docType: 'explain' },
           path: '/miniprogram/used',
-          component: () => import('@/../USED.md'),
+          component: () => import('@pages/used.vue'),
         },
         {
           title: '更新日志',
@@ -129,7 +129,7 @@ export default {
           component: () => import('@/navbar/README.md'),
         },
         {
-          title: 'SideBar 侧边导航',
+          title: 'SideBar 侧边栏',
           name: 'sidebar',
           meta: { docType: 'navigation' },
           path: '/miniprogram/components/side-bar',
@@ -177,7 +177,7 @@ export default {
           component: () => import('@/cascader/README.md'),
         },
         {
-          title: 'CheckBox 复选框',
+          title: 'CheckBox 多选框',
           name: 'checkbox',
           meta: { docType: 'form' },
           path: '/miniprogram/components/checkbox',
@@ -261,7 +261,7 @@ export default {
           component: () => import('@/tree-select/README.md'),
         },
         {
-          title: 'UpLoad 上传',
+          title: 'Upload 上传',
           name: 'upload',
           meta: { docType: 'form' },
           path: '/miniprogram/components/upload',
@@ -365,7 +365,7 @@ export default {
           component: () => import('@/skeleton/README.md'),
         },
         {
-          title: 'Sticky 吸顶容器',
+          title: 'Sticky 吸顶',
           name: 'sticky',
           meta: { docType: 'navigation' },
           path: '/miniprogram/components/sticky',
@@ -399,7 +399,7 @@ export default {
           component: () => import('@/action-sheet/README.md'),
         },
         {
-          title: 'Dialog 弹出框',
+          title: 'Dialog 对话框',
           name: 'dialog',
           meta: { docType: 'message' },
           path: '/miniprogram/components/dialog',