+++ src/avatar-group/props.ts
@@ -9,7 +9,7 @@ const props: TdAvatarGroupProps = {
   /** 图片之间的层叠关系，可选值：左侧图片在上和右侧图片在上 */
   cascading: {
     type: String,
-    value: 'right-up',
+    value: 'left-up',
   },
   /** 头像数量超出时，会出现一个头像折叠元素。该元素内容可自定义。默认为 `+N`。示例：`+5`，`...`, `更多` */
   collapseAvatar: {