+++ src/dialog/dialog.ts
@@ -104,8 +104,8 @@ export default class Dialog extends SuperComponent {
 
     onConfirm() {
       this.triggerEvent('confirm');
-      if (this._onComfirm) {
-        this._onComfirm();
+      if (this._onConfirm) {
+        this._onConfirm();
         this.close();
       }
     },
@@ -133,7 +133,7 @@ export default class Dialog extends SuperComponent {
       if (this.properties.closeOnOverlayClick) {
         this.triggerEvent('close', { trigger: 'overlay' });
       }
-      this.triggerEvent('overlayClick');
+      this.triggerEvent('overlay-click');
     },
 
     onActionTap(index: number) {