+++ src/input/input.ts
@@ -13,7 +13,6 @@ export default class Input extends SuperComponent {
   };
 
   externalClasses = [
-    'class',
     `${prefix}-class`,
     `${prefix}-class-prefix-icon`,
     `${prefix}-class-label`,