+++ src/message/_example/base/index.js
@@ -27,7 +27,12 @@ Component({
         offset: ['20rpx', 32],
         content: '这是一条带关闭的消息通知',
         duration: -1,
-        action: '按钮',
+        link: {
+          content: '按钮',
+          navigatorProps: {
+            url: '/page/xxx/xxx',
+          },
+        },
         closeBtn: true,
       });
     },
@@ -50,7 +55,12 @@ Component({
         icon: 'notification-filled',
         content: '这是一条带操作的消息通知',
         duration: -1,
-        action: '按钮',
+        link: {
+          content: '链接',
+          navigatorProps: {
+            url: '/page/xxx/xxx',
+          },
+        },
       });
     },
   },